+++ app/src/rx_video_app.c
@@ -562,7 +562,7 @@ static int app_rx_video_init(struct st_app_context* ctx, st_json_video_session_t
     ops.flags |= ST20_RX_FLAG_ENABLE_RTCP;
     ops.rtcp.nack_interval_us = 250;
   }
-  if (ctx->enable_timing_parser) ops.flags |= ST20_RX_FLAG_ENABLE_TIMING_PARSER;
+  if (ctx->enable_timing_parser) ops.flags |= ST20_RX_FLAG_TIMING_PARSER_STAT;
 
   st_pthread_mutex_init(&s->st20_wake_mutex, NULL);
   st_pthread_cond_init(&s->st20_wake_cond, NULL);